requirements.txt
fastapi[standard]
uvicorn
pydantic
pytest
sqlalchemy
dotenv
passlib[bcrypt]
pyjwt
# pytest-asyncio
# httpx
